// twitter/logic/models/trends.ts
// Generated by https://quicktype.io

export interface TrendsResponse {
  trends:     Trend[];
  as_of:      string;
  created_at: string;
  locations: Location[];

}

export interface Location {
  name:  string;
  woeid: number;
}

export interface Trend {
  name:             string;
  url:              string;
  promoted_content: null;
  query:            string;
  tweet_volume: number | null;
  tweets: any;
  azure_data: any;
  aws_data: any;
}